+++ tkcalendar/dateentry.py
@@ -31,6 +31,7 @@ try:
 except ImportError:
     import Tkinter as tk
     import ttk
+
 from tkcalendar.calendar_ import Calendar