+++ docs/changelog.rst
@@ -3,6 +3,25 @@ Changelog
 
 .. currentmodule:: tkcalendar
 
+tkcalendar 1.5.0
+----------------
+
+.. rubric:: New Calendar options
+
+- *disabledforeground* and *disabledbackground*: colors of calendar border and
+  month/year name in disabled state
+  
+- *maxdate* and *mindate*: set an allowed date range for date selection
+
+.. rubric:: New features
+
+- :meth:`Calendar.see` method: make sure given date is visible
+
+.. rubric:: Bug fixes
+
+- Make :meth:`Calendar.selection_clear` actually clear the selection
+
+
 tkcalendar 1.4.0
 ----------------